refactor: Deprecate legacy layout
