+++ packages/react-utils/src/hooks/index.ts
@@ -10,6 +10,7 @@ export * from './useDebounce'
 export * from './useDebounceCallback'
 export * from './useDocumentTitle'
 export * from './useElementSize'
+export * from './useElementTopOffset'
 export * from './useEventHandler'
 export * from './useExpectSameValueReference'
 export * from './useForceRender'
