+++ packages/ui/src/components/index.css
@@ -54,4 +54,3 @@
 
 /* Complex components */
 @import "./Editor";
-@import "./Layout";
