+++ packages/ui/src/components/index.ts
@@ -13,18 +13,18 @@ export * from './DimensionSwitcher'
 export * from './Divider'
 export * from './Dropdown'
 export * from './Editor'
+export * from './EmptyStateContainer'
 export * from './ErrorList'
 export * from './Forms'
 export * from './Grid'
 export * from './Icon'
-export * from './Providers'
-export * from './Layout'
 export * from './Logo'
 export * from './Menu'
 export * from './Message'
 export * from './Portal'
 export * from './PreventCloseContext'
 export * from './ProgressBar'
+export * from './Providers'
 export * from './RepeaterItemContainer'
 export * from './SaveButton'
 export * from './SeamlessDropdown'
@@ -35,7 +35,6 @@ export * from './Spinner'
 export * from './SpinnerContainer'
 export * from './Stack'
 export * from './StyleProvider'
-export * from './EmptyStateContainer'
 export * from './Table'
 export * from './Tabs'
 export * from './Tag'
