+++ packages/ui/src/deprecated/Layout/index.ts
@@ -1,8 +1,8 @@
+export { useElementTopOffset } from '@contember/react-utils'
 export * from './Layout'
 export * from './LayoutChrome'
 export * from './LayoutPage'
 export * from './LayoutPageAside'
 export * from './LayoutPageContent'
 export * from './LayoutPageStickyContainer'
-export * from './useElementTopOffset'
 export type { ThemeScheme } from './Types'
