+++ packages/ui/src/deprecated/index.css
@@ -1,1 +1,2 @@
 /* DEPRECATED */
+@import "./Layout/index.css";
