+++ packages/ui/src/deprecated/index.ts
@@ -4,6 +4,7 @@ export default {}
 
 export * from './ContainerSpinner'
 export * from './DropdownContentContainerContext'
+export * from './Layout'
 export * from './Portal'
 export * from './toSchemeClass'
 export * from './toThemeClass'
